Run jobs in separate threads, and debug logging.
Previously, since the jobs were run in the same thread, exceptions
delivered to the worker thread would be caught by the exception handler
for the job. In practice, this may not be a big deal, but it makes
it impossible to test, as I can't cause the worker thread to fail
independently!