Monitor every 1/10th job timeout, not 10ms.
In general, the job timeout is going to be quite long, and there is no
reason to be checking with the monitor so frequently. It could probably
be even longer (really, if the job timeout is 10 minutes, and you add
it back to the queue after 12 or 13, it doesn't matter).

Also, slow down the multiple monitors test even more...